+++ src/App.js
@@ -3,6 +3,7 @@ import { createBrowserRouter, Navigate, RouterProvider } from "react-router-dom"
 import Root from "./Root.js";
 import RouteError from "./components/RouteError/RouteError.jsx";
 import Login from "./components/Login/Login.jsx";
+import Home from "./components/Main/Home/Home.jsx";
 import EmployeeContextProvider from "./store/employee-context.jsx";
 import "./App.css";
 
@@ -20,7 +21,8 @@ const router = createBrowserRouter([
     element: <EmployeeContextProvider><Root /></EmployeeContextProvider>,
     errorElement: <RouteError root={true}/>,
     children: [
-      // { index: true, element: <Navigate to="/metrics" replace /> },
+      { index: true, element: <Navigate to={'/home'} replace/>},
+      { path: '/home', element: <Home>starting page</Home>, errorElement: <RouteError />},
       { path: '/metrics', element: <Suspense><Metrics /></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Metrics/Metrics.jsx').then(module => module.thresholdsLoader())},
       { path: '/selector', element: <Suspense><SelectorContextProvider><Selector /></SelectorContextProvider></Suspense>, errorElement: <RouteError />, loader: () => import('./components/Main/Selector/Selector.jsx').then(module => module.filtersLoader())},
       { path: '/company', element: <Suspense><CompanyContextProvider><Company /></CompanyContextProvider></Suspense>},