+++ CHANGELOG.md
@@ -3,6 +3,7 @@
 * [@BartoszDziurowicz](https://github.com/BartekDziurowicz/carbon-tracker-web/pull/27)
 * Implement logic to handle extended  by LDAP calls and objects
 * Create login page and login handling using base auth
+* Handle location/path when manually changed
 
 0.36 (2024-10-19)
 ===============